Move device classes to own files
