+++ freeathome/fah/pfreeathome.py
@@ -20,37 +20,18 @@ from slixmpp.plugins.xep_0009.stanza.RPC import RPCQuery, MethodCall, MethodResp
 from slixmpp.plugins.xep_0060.stanza.pubsub_event import Event, EventItems, EventItem
 from slixmpp.exceptions import IqError
 from slixmpp import Iq
+
+from .devices.fah_light import FahLight
+from .devices.fah_binary_sensor import FahBinarySensor
+from .devices.fah_thermostat import FahThermostat
+from .devices.fah_light_scene import FahLightScene
+from .devices.fah_cover import FahCover
+
 from .const import (
     FUNCTION_IDS_SENSOR_UNIT,
-    FUNCTION_IDS_SWITCHING_ACTUATOR,
-    FUNCTION_IDS_DIMMING_ACTUATOR,
-    FUNCTION_IDS_BLIND_ACTUATOR,
-    FUNCTION_IDS_SCENE,
-    FUNCTION_IDS_ROOM_TEMPERATURE_CONTROLLER,
-    FUNCTION_IDS_MOVEMENT_DETECTOR,
     NAME_IDS_TO_BINARY_SENSOR_SUFFIX,
-    PID_SWITCH_ON_OFF,
-    PID_ABSOLUTE_SET_VALUE,
-    PID_INFO_ON_OFF,
-    PID_INFO_ACTUAL_DIMMING_VALUE,
-    PID_MOVE_UP_DOWN,
-    PID_ADJUST_UP_DOWN,
-    PID_SET_ABSOLUTE_POSITION_BLINDS,
-    PID_FORCE_POSITION_BLIND,
-    PID_INFO_MOVE_UP_DOWN,
-    PID_CURRENT_ABSOLUTE_POSITION_BLINDS_PERCENTAGE,
-    PID_FORCE_POSITION_INFO,
-    PID_ECO_MODE_ON_OFF_REQUEST,
-    PID_CONTROLLER_ON_OFF_REQUEST,
-    PID_ABSOLUTE_SETPOINT_TEMPERATURE,
-    PID_SET_VALUE_TEMPERATURE,
-    PID_CONTROLLER_ON_OFF,
-    PID_STATUS_INDICATION,
-    PID_MEASURED_TEMPERATURE,
-    PID_HEATING_DEMAND,
-    PID_PRESENCE,
-    PID_SCENE_CONTROL,
     )
+
 from .messagereader import MessageReader
 from .settings import SettingsFah
 from .saslhandler import SaslHandler
@@ -88,387 +69,6 @@ def message2py(mes):
     for data in mes.xml.findall('{%s}data' % namespace):
         vals.append(data.text)
     return vals
-
-
-class FahDevice:
-    """ Free@Home base object """
-
-    def __init__(self, client, device_info, serialnumber, channel_id, name, datapoints={}, device_updated_cb=None):
-        self._device_info = device_info
-        self._serialnumber = serialnumber
-        self._channel_id = channel_id
-        self._name = name
-        self._client = client
-        self._device_updated_cbs = []
-        self._datapoints = datapoints
-        if device_updated_cb is not None:
-            self.register_device_updated_cb(device_updated_cb)
-
-    def register_device_updated_cb(self, device_updated_cb):
-        """Register device updated callback."""
-        self._device_updated_cbs.append(device_updated_cb)
-
-    def unregister_device_cb(self, device_updated_cb):
-        """Unregister device updated callback."""
-        self._device_updated_cbs.remove(device_updated_cb)
-
-    async def after_update(self):
-        """Execute callbacks after internal state has been changed."""
-        for device_updated_cb in self._device_updated_cbs:
-            await device_updated_cb(self)
-
-    @property
-    def serialnumber(self):
-        """ return the serial number """
-        return self._serialnumber
-
-    @property
-    def channel_id(self):
-        """Return channel id"""
-        return self._channel_id
-
-    @property
-    def name(self):
-        """ return the name of the device   """
-        return self._name
-
-    @property
-    def client(self):
-        """ return the Client object """
-        return self._client
-
-    @property
-    def device_info(self):
-        """Return device info."""
-        return self._device_info
-
-    @property
-    def lookup_key(self):
-        """Return device lookup key"""
-        return self.serialnumber + "/" + self.channel_id
-
-# TODO: Use FahSensor for movement detector lux sensor
-class FahSensor(FahDevice):
-    """ Free@Home sensor object """
-    state = None     
-    output_device = None    
-
-    def __init__(self, client, device_info, serialnumber, channel_id, name, sensor_type, state, output_device):
-        FahDevice.__init__(self, client, device_info, serialnumber, channel_id, name)
-        self.type = sensor_type
-        self.state = state        
-        self.output_device = output_device
-
-class FahBinarySensor(FahDevice):
-    """Free@Home binary object """
-    state = None
-    output_device = None
-
-    def pairing_ids(function_id=None):
-        if function_id in FUNCTION_IDS_SENSOR_UNIT:
-            return {
-                    "inputs": [],
-                    "outputs": [
-                        PID_SWITCH_ON_OFF,
-                        ]
-                    }
-
-        elif function_id in FUNCTION_IDS_MOVEMENT_DETECTOR:
-            return {
-                    "inputs": [],
-                    "outputs": [
-                        PID_PRESENCE,
-                        ]
-                    }
-
-    def update_datapoint(self, dp, value):
-        """Receive updated datapoint."""
-        if self._datapoints.get(PID_SWITCH_ON_OFF) == dp or \
-                self._datapoints.get(PID_PRESENCE) == dp:
-            self.state = value
-            LOG.debug("binary sensor device %s dp %s state %s", self.lookup_key, dp, value)
-
-
-class FahLock(FahDevice):
-    """" Free@home lock controll in 7 inch panel """
-    state = None
-
-    def __init__(self, client, device_info, serialnumber, channel_id, name, state=False):
-        FahDevice.__init__(self, client, device_info, serialnumber, channel_id, name)
-        self.state = state        
-
-    async def lock(self):
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, 'idp0000', '0')
-    
-    async def unlock(self):
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, 'idp0000', '1')
-
-class FahThermostat(FahDevice):
-    """Free@Home thermostat """
-    current_temperature = None
-    current_actuator = None
-    target_temperature = None
-
-    def pairing_ids(function_id=None):
-        if function_id in FUNCTION_IDS_ROOM_TEMPERATURE_CONTROLLER:
-            return {
-                    "inputs": [
-                        PID_ECO_MODE_ON_OFF_REQUEST,
-                        PID_CONTROLLER_ON_OFF_REQUEST,
-                        PID_ABSOLUTE_SETPOINT_TEMPERATURE,
-                        ],
-                    "outputs": [
-                        PID_SET_VALUE_TEMPERATURE,
-                        PID_CONTROLLER_ON_OFF,
-                        PID_STATUS_INDICATION,
-                        PID_MEASURED_TEMPERATURE,
-                        PID_HEATING_DEMAND,
-                        ]
-                    }
-
-    async def turn_on(self):
-        """ Turn the thermostat on   """
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_ECO_MODE_ON_OFF_REQUEST], '0')
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_CONTROLLER_ON_OFF_REQUEST], '1')
-
-    async def turn_off(self):
-        """ Turn the thermostat off   """
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_CONTROLLER_ON_OFF_REQUEST], '0')
-
-    async def eco_mode(self):
-        """ Put the thermostat in eco mode   """
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_ECO_MODE_ON_OFF_REQUEST], '2')
-
-    async def set_target_temperature(self, temperature):
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_ABSOLUTE_SETPOINT_TEMPERATURE], '%.2f' % temperature)
-
-    @property
-    def state(self):
-        return self._state
-
-    @state.setter
-    def state(self, state):
-        self._state = state == '1'
-
-    @property
-    def ecomode(self):
-        return self._eco_mode
-
-    @ecomode.setter
-    def ecomode(self, eco_mode):
-        self._eco_mode = int(eco_mode) & 0x04 == 0x04
-
-    def update_datapoint(self, dp, value):
-        """Receive updated datapoint."""
-        if self._datapoints.get(PID_SET_VALUE_TEMPERATURE) == dp:
-            self.target_temperature = value
-            LOG.info("thermostat device %s target temp is %s", self.lookup_key, value)
-
-        elif self._datapoints.get(PID_CONTROLLER_ON_OFF) == dp:
-            self.state = value
-            LOG.info("thermostat device %s state is %s", self.lookup_key, value)
-
-        elif self._datapoints.get(PID_STATUS_INDICATION) == dp:
-            self.ecomode = value
-            LOG.info("thermostat device %s eco mode is %s", self.lookup_key, value)
-
-        elif self._datapoints.get(PID_MEASURED_TEMPERATURE) == dp:
-            self.current_temperature = value
-            LOG.info("thermostat device %s current temp is %s", self.lookup_key, value)
-
-        elif self._datapoints.get(PID_HEATING_DEMAND) == dp:
-            self.current_actuator = value
-            LOG.info("thermostat device %s current heating actuator state is %s", self.lookup_key, value)
-
-
-class FahLight(FahDevice):
-    """ Free@Home light object   """
-    state = None
-    light_type = None
-    brightness = None
-
-    def pairing_ids(function_id=None):
-        if function_id in FUNCTION_IDS_SWITCHING_ACTUATOR:
-            return {
-                    "inputs": [
-                        PID_SWITCH_ON_OFF,
-                        ],
-                    "outputs": [
-                        PID_INFO_ON_OFF,
-                        ]
-                    }
-        elif function_id in FUNCTION_IDS_DIMMING_ACTUATOR:
-            return {
-                    "inputs": [
-                        PID_SWITCH_ON_OFF,
-                        PID_ABSOLUTE_SET_VALUE,
-                        ],
-                    "outputs": [
-                        PID_INFO_ON_OFF,
-                        PID_INFO_ACTUAL_DIMMING_VALUE,
-                        ]
-                    }
-
-    async def turn_on(self):
-        """ Turn the light on   """
-        oldstate = self.state
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_SWITCH_ON_OFF], '1')
-        self.state = True
-
-        if self.is_dimmer() \
-                and ((oldstate != self.state and int(self.brightness) > 0) or (oldstate == self.state)):
-            await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_ABSOLUTE_SET_VALUE], str(self.brightness))
-
-    async def turn_off(self):
-        """ Turn the light off   """
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, self._datapoints[PID_SWITCH_ON_OFF], '0')
-        self.state = False
-
-    def set_brightness(self, brightness):
-        """ Set the brightness of the light  """
-        if self.is_dimmer():
-            self.brightness = brightness
-
-    def get_brightness(self):
-        """ Return the brightness of the light  """
-        return self.brightness
-
-    def is_on(self):
-        """ Return the state of the light   """
-        return self.state
-
-    def is_dimmer(self):
-        """Return true if device is a dimmer"""
-        return PID_ABSOLUTE_SET_VALUE in self._datapoints
-
-    def update_datapoint(self, dp, value):
-        """Receive updated datapoint."""
-        if PID_INFO_ON_OFF in self._datapoints and self._datapoints[PID_INFO_ON_OFF] == dp:
-            self.state = (value == '1')
-            LOG.info("light device %s (%s) is %s",
-                     self.name, self.lookup_key, self.state)
-
-        elif PID_INFO_ACTUAL_DIMMING_VALUE in self._datapoints and self._datapoints[PID_INFO_ACTUAL_DIMMING_VALUE] == dp:
-            self.brightness = value
-            LOG.info("light device %s (%s) brightness %s",
-                     self.name, self.lookup_key,
-                     self.brightness)
-
-# TODO: Add a FahTimerProfile class for timer profiles (function ID 4A00)
-# Timer profiles may be switched on and off, so they are not exactly
-# like scenes, but more like switches
-
-class FahLightScene(FahDevice):
-    """ Free@home scene   """
-
-    def pairing_ids(function_id=None):
-        if function_id in FUNCTION_IDS_SCENE:
-            return {
-                    "inputs": [],
-                    "outputs": [
-                        PID_SCENE_CONTROL,
-                        ]
-                    }
-
-    async def activate(self):
-        """ Activate the scene   """
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, 'odp0000', '1')
-
-    def update_datapoint(self, dp, value):
-        """Receive updated datapoint."""
-        if self._datapoints.get(PID_SCENE_CONTROL) == dp:
-            self.state = value
-            LOG.info("scene %s (%s) is %s", self.name, self.lookup_key, self.state)
-
-
-class FahCover(FahDevice):
-    """ Free@Home cover device
-    In freeathome the value 100 indicates that the cover is fully closed
-    In home assistant the value 100 indicates that the cover is fully open
-    """
-    state = None
-    position = None
-    forced_position = None
-
-    def pairing_ids(function_id=None):
-        # TODO: Determine from function ID which pairing IDs are relevant
-        # E.g. Slats -> Add slats data points
-        if function_id in FUNCTION_IDS_BLIND_ACTUATOR:
-            return {
-                    "inputs": [
-                        PID_MOVE_UP_DOWN,
-                        PID_ADJUST_UP_DOWN,
-                        PID_SET_ABSOLUTE_POSITION_BLINDS,
-                        PID_FORCE_POSITION_BLIND,
-                        ],
-                    "outputs": [
-                        PID_INFO_MOVE_UP_DOWN,
-                        PID_CURRENT_ABSOLUTE_POSITION_BLINDS_PERCENTAGE,
-                        PID_FORCE_POSITION_INFO,
-                        ]
-                    }
-
-    def is_cover_closed(self):
-        """ Return if the cover is closed   """
-        return int(self.position) == 0
-
-    def is_cover_opening(self):
-        """ Return is the cover is opening   """
-        return self.state == '2'
-
-    def is_cover_closing(self):
-        """ Return if the cover is closing   """
-        return self.state == '3'
-
-    def get_cover_position(self):
-        """ Return the cover position """
-        return int(self.position)
-
-    def get_forced_cover_position(self):
-        """Return forced cover position."""
-        return int(self.forced_position)
-
-    async def set_cover_position(self, position):
-        """ Set the cover position  """
-        dp = self._datapoints[PID_SET_ABSOLUTE_POSITION_BLINDS]
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, dp, str(abs(100 - position)))
-
-    async def set_forced_cover_position(self, forced_position):
-        """Set forced cover position."""
-        dp = self._datapoints[PID_FORCE_POSITION_BLIND]
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, dp, str(forced_position))
-
-    async def open_cover(self):
-        """ Open the cover   """
-        dp = self._datapoints[PID_MOVE_UP_DOWN]
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, dp, '0')
-
-    async def close_cover(self):
-        """ Close the cover   """
-        dp = self._datapoints[PID_MOVE_UP_DOWN]
-        await self.client.set_datapoint(self.serialnumber, self.channel_id, dp, '1')
-
-    async def stop_cover(self):
-        """ Stop the cover, only if it is moving """
-        if (self.state == '2') or (self.state == '3'):
-            dp = self._datapoints[PID_ADJUST_UP_DOWN]
-            await self.client.set_datapoint(self.serialnumber, self.channel_id, dp, '1')
-
-    def update_datapoint(self, dp, value):
-        """Receive updated datapoint."""
-        if self._datapoints.get(PID_INFO_MOVE_UP_DOWN) == dp:
-            self.state = value
-            LOG.info("cover device %s (%s) is %s",
-                     self.name, self.lookup_key, self.state)
-
-        elif self._datapoints.get(PID_CURRENT_ABSOLUTE_POSITION_BLINDS_PERCENTAGE) == dp:
-            self.position = str(abs(100 - int(float(value))))
-            LOG.info("cover device %s (%s) position %s",
-                     self.name, self.lookup_key,
-                     self.position)
-
-        elif self._datapoints.get(PID_FORCE_POSITION_INFO) == dp:
-            self.forced_position = value
 
 
 def get_room_names(xmlroot):
